Spacing issues, group formatting
I cleaned up the spacing for the groups page so it can handle different resolutions, though text scaling is still an issue. I also implemented expandable groups on the my groups and discover pages, though the chat buttons on them don't currently go anywhere (as frontend for chats is not yet implemented).